refactor(desktop): keep project hooks within size limits
Move relay project fetching into a focused module so the startup optimization remains maintainable and passes the repository file-size gate.
