+++ desktop/src/features/projects/hooks.ts
@@ -3,7 +3,6 @@ import * as React from "react";
 
 import { relayClient } from "@/shared/api/relayClient";
 import { getRelaySelf } from "@/features/moderation/lib/relaySelf";
-import { getCachedRelayOrigin } from "@/shared/lib/mediaUrl";
 import { signRelayEvent } from "@/shared/api/tauri";
 import { getIdentity } from "@/shared/api/tauriIdentity";
 import {
@@ -64,15 +63,11 @@ import {
   type Project,
   type Repository,
 } from "./projectModels";
+import { fetchProjectHomeForChannel, fetchProjects } from "./projectFetch";
 import { persistProjectSnapshot, PROJECTS_QUERY_KEY } from "./projectSnapshot";
-import {
-  buildProjectHomeFromFetcher,
-  buildProjectsFromFetcher,
-  type FetchProjectEventsExhaustively,
-  fetchProjectEventsExhaustively,
-} from "./projectEnumeration";
 import { projectMatchesRouteId } from "./projectRoutes";
 
+export { fetchProjects } from "./projectFetch";
 export type {
   Project,
   ProjectIssue,
@@ -82,7 +77,6 @@ export type {
 };
 export type ProjectPullRequestCommentDecision = "request-changes";
 
-const HIDDEN_PROJECT_CARDS_KEY = "buzz.projects.hidden-cards.v1";
 export type RepoState = {
   branches: Array<{ name: string; commit: string }>;
   tags: Array<{ name: string; commit: string }>;
@@ -131,23 +125,6 @@ export type ProjectIssueListItem = {
   issue: ProjectIssue;
 };
 
-function readHiddenProjectCards(): string[] {
-  if (typeof window === "undefined") {
-    return [];
-  }
-
-  try {
-    const parsed = JSON.parse(
-      window.localStorage.getItem(HIDDEN_PROJECT_CARDS_KEY) ?? "[]",
-    );
-    return Array.isArray(parsed)
-      ? parsed.filter((item): item is string => typeof item === "string")
-      : [];
-  } catch {
-    return [];
-  }
-}
-
 /**
  * Converts a kind:30617 repo announcement into a `Project`.
  *
@@ -165,46 +142,6 @@ export function eventToProject(
     throw new Error("Invalid repository announcement.");
   }
   return repository;
-}
-
-export async function fetchProjects(
-  fetchExhaustively?: FetchProjectEventsExhaustively,
-  signal?: AbortSignal,
-): Promise<Project[]> {
-  // Delegates to `buildProjectsFromFetcher` in `projectEnumeration.ts`, which
-  // is the pure, Tauri-free core of this operation. Its javadoc explains
-  // fail-closed tombstones and NIP-OA owner-deletion suppression.
-  const viewerPubkey = await getIdentity()
-    .then((identity) => identity.pubkey)
-    .catch(() => undefined);
-  const fetcher: FetchProjectEventsExhaustively =
-    fetchExhaustively ??
-    ((kinds, extraFilter) =>
-      fetchProjectEventsExhaustively(kinds, extraFilter, undefined, signal));
-  return buildProjectsFromFetcher(fetcher, {
-    relayOrigin: getCachedRelayOrigin(),
-    hiddenAddresses: new Set(readHiddenProjectCards()),
-    viewerPubkey,
-  });
-}
-
-async function fetchProjectHomeForChannel(
-  channelId: string,
-  signal?: AbortSignal,
-): Promise<Project | null> {
-  const viewerPubkey = await getIdentity()
-    .then((identity) => identity.pubkey)
-    .catch(() => undefined);
-  return buildProjectHomeFromFetcher(
-    (kinds, extraFilter) =>
-      fetchProjectEventsExhaustively(kinds, extraFilter, undefined, signal),
-    channelId,
-    {
-      relayOrigin: getCachedRelayOrigin(),
-      hiddenAddresses: new Set(readHiddenProjectCards()),
-      viewerPubkey,
-    },
-  );
 }
 
 function eventToRepoState(event: RelayEvent): RepoState {
