+++ desktop/src/features/projects/projectFetch.ts
@@ -1,0 +1,70 @@
+import { getCachedRelayOrigin } from "@/shared/lib/mediaUrl";
+import { getIdentity } from "@/shared/api/tauriIdentity";
+import {
+  buildProjectHomeFromFetcher,
+  buildProjectsFromFetcher,
+  type FetchProjectEventsExhaustively,
+  fetchProjectEventsExhaustively,
+} from "./projectEnumeration";
+import type { Project } from "./projectModels";
+
+const HIDDEN_PROJECT_CARDS_KEY = "buzz.projects.hidden-cards.v1";
+
+function readHiddenProjectCards(): string[] {
+  if (typeof window === "undefined") {
+    return [];
+  }
+
+  try {
+    const parsed = JSON.parse(
+      window.localStorage.getItem(HIDDEN_PROJECT_CARDS_KEY) ?? "[]",
+    );
+    return Array.isArray(parsed)
+      ? parsed.filter((item): item is string => typeof item === "string")
+      : [];
+  } catch {
+    return [];
+  }
+}
+
+/** Enumerates the projects visible to the current relay identity. */
+export async function fetchProjects(
+  fetchExhaustively?: FetchProjectEventsExhaustively,
+  signal?: AbortSignal,
+): Promise<Project[]> {
+  // Delegates to `buildProjectsFromFetcher` in `projectEnumeration.ts`, which
+  // is the pure, Tauri-free core of this operation. Its javadoc explains
+  // fail-closed tombstones and NIP-OA owner-deletion suppression.
+  const viewerPubkey = await getIdentity()
+    .then((identity) => identity.pubkey)
+    .catch(() => undefined);
+  const fetcher: FetchProjectEventsExhaustively =
+    fetchExhaustively ??
+    ((kinds, extraFilter) =>
+      fetchProjectEventsExhaustively(kinds, extraFilter, undefined, signal));
+  return buildProjectsFromFetcher(fetcher, {
+    relayOrigin: getCachedRelayOrigin(),
+    hiddenAddresses: new Set(readHiddenProjectCards()),
+    viewerPubkey,
+  });
+}
+
+/** Resolves the active channel's project home with a scoped relay query. */
+export async function fetchProjectHomeForChannel(
+  channelId: string,
+  signal?: AbortSignal,
+): Promise<Project | null> {
+  const viewerPubkey = await getIdentity()
+    .then((identity) => identity.pubkey)
+    .catch(() => undefined);
+  return buildProjectHomeFromFetcher(
+    (kinds, extraFilter) =>
+      fetchProjectEventsExhaustively(kinds, extraFilter, undefined, signal),
+    channelId,
+    {
+      relayOrigin: getCachedRelayOrigin(),
+      hiddenAddresses: new Set(readHiddenProjectCards()),
+      viewerPubkey,
+    },
+  );
+}
